add on-disk cache to significantly speed up process
This is a relatively naive implementation that tries to use an on-disk cache.

The on-disk cache is persisted when a build has finished. buildEnd fires when:

* the dev server is gracefully shut down
* the dev server restarts
* the production build process exits

The on-disk cache is used to populate the already existing in-memory
Map, which was intended by the module authors to be used as a cache.
This cache contains mappings from various kinds of module specifiers to
the info that is needed to accurately resolve a module. In some cases,
we get a null result when resolving an id because the target isn't
something that we can resolve in this plugin. These cases were ignored
in the upstream code and not put into the cache, causing unnecessary and
time-costly shell invocations of `deno info --json ${id}`.

In general, this change aims to put all results into the cache such that
whenever it's hit we either:
* get what we need for letting Vite resolve the id, or
* we get the null result we would get in all cases but can now skip the
  costly `deno info --json` check that was occuring.

This change has brought development server runtime before rendering
anything on screen from multiple minutes (sometimes as much as 5
minutes) to a few seconds.  Production builds also run quick, clocking
in at ~24s for a large codebase when the cache exists.

Likely still needs more work structure-wise but it's a significant
quality of life improvement on my machine already.

One thing to note: the server restarts in our project can be a bit wonky
and may / may not invalidate the on-disk cache by causing a write? Not
sure yet, and could maybe be mitigated by setting hmr: false in Vite's
config.